Version 3.0: major upgrade
Basically a re-write of the basic method, this version stops using
the device ID as a hardcoded variable and starts dynamically connecting
to the device the user specifies, using transport_id.

This way, the way the device is connected is irrelevant, because the product of
the device stays the same. This means ADB Over WIFI is supported.

This also means that the user can bypass the alias scripts and directly connect
to any device by specifying its product variable, e.g.:
`adb-multi bullhead shell`

Notes: wait-for-device was removed because it can't work, if the device is not
connected, there is no transport_id or ID to pass to it.